app/requirements.txt
Flask
flask_mysql
yaml
